requirements.txt
alembic==1.10.3
Babel==2.12.1
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
flask-msearch==0.2.9.2
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.0.3
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
python-dotenv==1.0.0
python-slugify==8.0.1
speaklater==1.3
SQLAlchemy==2.0.5.post1
text-unidecode==1.3
typing_extensions==4.5.0
Werkzeug==2.2.3
WTForms==3.0.1
